% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{[.models}
\alias{[.models}
\title{Extracts fitted models from a "models" object. Supports indexation
by formula and integers.}
\usage{
\method{[}{models}(models, x)
}
\arguments{
\item{x}{An object of class "subformula" or integer vector.}
}
\value{
A models object.
}
\description{
Extracts fitted models from a "models" object. Supports indexation
by formula and integers.
}
