% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{[.subformulas}
\alias{[.subformulas}
\title{Extracts fitted models from a "models" object. Supports indexation
integers.}
\usage{
\method{[}{subformulas}(forms, x)
}
\arguments{
\item{x}{An object of class "subformula" or integer vector.}
}
\value{
A subformulas object.
}
\description{
Extracts fitted models from a "models" object. Supports indexation
integers.
}
