% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexation.R
\name{subformula_indices}
\alias{subformula_indices}
\title{Takes to subformlas objects "subs1" and "subs2", and finds the indices
"indices" such that subs1[indices] = subs2. If such an indices vector
does not exist, it throws an error.}
\usage{
subformula_indices(subs1, subs2)
}
\arguments{
\item{subs1}{An object of class "subformula". Should contain subs2.}

\item{subs2}{An object of class "subformula".}
}
\value{
An integer vector of indices satisfying subs1[indices] = subs2.
}
\description{
Takes to subformlas objects "subs1" and "subs2", and finds the indices
"indices" such that subs1[indices] = subs2. If such an indices vector
does not exist, it throws an error.
}
