% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subformulas.R
\name{subformulas}
\alias{subformulas}
\title{Find subformulas of a formula with protected covariates.}
\usage{
subformulas(formula, protected = NULL, as_formula = TRUE,
  keep_interactions = TRUE, intercept_protected = TRUE, data = NULL)
}
\arguments{
\item{formula}{A formula or string specifying a formula.}

\item{protected}{A vector or formula specifying which, if any,
covariates are protected. They will appear in all computed
subformulas.}

\item{as_formula}{Bool specifying whether the function returns
objects of class formula or string.}

\item{keep_interactions}{If true, interaction terms will only appear
together with their parents: For x:y to appear, x and y must also appear.}

\item{data}{Optional supplied data. Is used to fill out formulas as in
"y ~ .".}
}
\value{
A list of class "formulas".
}
\description{
Find subformulas of a formula with protected covariates.
}
\details{
If the supplied formula includes the term "0" or "-1", none
of the subformulas will include the intercept. Otherwise, the intercept
will be interpreted as being protected.
}
\examples{
subformulas(z ~ x + y)
subformulas(y ~ x + y + y^2, protected = ~ x)
subformulas(y ~ x + y + t + I(t^2), protected = c("x","I(t^2)"))
}
