% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapply.R
\name{fapply}
\alias{fapply}
\title{Apply Formulas to a Model}
\usage{
fapply(formulas, model, ...)
}
\arguments{
\item{formulas}{a list of formulas or objects coercible to formula by
\code{\link[stats]{stats::as.formula}}.}

\item{model}{a function taking a \code{formula} as its first argument.}

\item{...}{additional arguments to be passed to \code{model}.}
}
\value{
\code{fapply} returns a list of evaluated function calls.
}
\description{
This is a member of the \code{apply} family. It is similar to
\code{lapply}, but handles the \code{call} slightly differently.
This makes the output prettier.
}
\examples{
formulas = subformulas(mpg ~ cyl + disp, protected = ~ cyl)
fapply(formulas, lm, data = mtcars) # Pretty output.
lapply(formulas, lm, data = mtcars) # Less pretty output.
}
