% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapply.R
\name{fapply}
\alias{fapply}
\title{Apply formulas to a model function}
\usage{
fapply(formulas, model, data, ...)
}
\arguments{
\item{formulas}{A subformulas object.}

\item{model}{A model function. Should take a formula as its first argument.}

\item{data}{An optional data frame containing the covariates and response
referenced in the formulas supplied. The data argument overrides the environments}

\item{...}{Ellipses are passed to the model function.}
}
\value{
A list of class "formulas".
}
\description{
Apply formulas to a model function
}
\details{
If the supplied formula includes the term "0" or "-1", none
of the subformulas will include the intercept. Otherwise, the intercept
will be interpreted as being protected.
}
\examples{
data(mtcars)
forms = subformulas(mpg ~ cyl + disp + hp + drat,
                    protected = ~ cyl)
mods = fapply(formulas = forms,
              model = lm,
              data = mtcars)
}
