% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapply.R
\name{fapply}
\alias{fapply}
\title{Apply formulas to a model function}
\usage{
fapply(formulas, model, ...)
}
\arguments{
\item{formulas}{an object of class "\code{subformulas}".}

\item{model}{a function taking a \code{formula} as its first argument.}

\item{...}{additional arguments to be passed to the \code{model}.}
}
\value{
A list of evaluated function calls.
}
\description{
Apply formulas to a model function
}
\examples{
formulas = subformulas(mpg ~ cyl + disp + hp + drat, protected = ~ cyl)
models = fapply(formulas = formulas, model = lm, data = mtcars)
}
